# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is FundooNotes Api",
        "version": "1.0.0",
        "title": "FundooNotes API"
    },
    "host": "localhost:3000",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/labels/{id}":{
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "Find Label By Id..!!",
                    "type": "string"
                }
            ],
            "put": {
                "tags": [
                    "FundooNote: Update Labels"
                ],
                "summary": "Update An Existing Label..!!",
                "description": "Find Label By Id And Update..!!",
                "operationId": "updateLabel",
                "parameters": [
                    {
                        "name": "labels",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type":"object",
                            "properties": {
                                "label":{
                                    "type":"string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200":{
                        "description": "Label Get Updated Successfully...!!"
                    },
                    "404":{
                        "description": "Failed To Update Label...!!!"
                    }
                },
                "security": [
                    {
                        "JWT":[]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "FundooNotes: Delete Label"
                ],
                "summary": "Delete A Label",
                "description": "Delete A Label By Its ID..!",
                "operationId": "deleteLabel",
                "responses": {
                    "200":{
                        "description": "Label Get deleted Successully...!!!"
                    },
                    "400":{
                        "description": "Failed To Delete Label...!!!"
                    }
                },
                "security": [
                    {
                        "JWT":[]
                    }
                ]
                
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}